refactor(crypto): move encrypted-body.dto to crypto module
Restructure the codebase by moving EncryptedBodyDto to the crypto module where it logically belongs. This improves code organization and maintainability by keeping related cryptographic components together.
