+++ src/domain/crypto/dtos/encrypted-body.dto.ts
@@ -2,8 +2,8 @@ import {
   IEncryptedBody,
   IEncryption,
   ISignature,
-} from "../interfaces/encrypted-body.interfaces";
-import { TRawJson } from "../shared/interfaces/general.interfaces";
+} from "../../interfaces/encrypted-body.interfaces";
+import { TRawJson } from "../../shared/interfaces/general.interfaces";
 
 import { ERRORS } from "@/config/strings/global.strings.json";
