+++ src/infrastructure/validators/encrypted-body.validator.ts
@@ -1,7 +1,7 @@
 import { z } from "zod";
 import { BadRequestError } from "@/domain/errors/bad-request-error";
 import type { TRawJson } from "@/domain/shared/interfaces/general.interfaces";
-import { EncryptedBodyDto } from "@/domain/dtos/encrypted-body.dto";
+import { EncryptedBodyDto } from "@/domain/crypto/dtos/encrypted-body.dto";
 import { processValidationError } from "./utils/processError.validator";
 
 // Base64url regex: URL-safe base64 without padding
